fix(signature): demo拆解与规范
